Fix missing "await"
I rarely do this, so I had forgotten the repercussions. I decided not
to wait on the insert promise to finish, since whether it succeeded
or failed I didn't really care. However, this means that if the promise
rejects without a catch it will crash.

It's no big deal to just wait, and technically better to show the
user a failure message if their vote wasn't saved anyway.